refactor App structure
