+++ client/src/contexts/QueryContext.js
@@ -1,0 +1,3 @@
+import React from 'react';
+
+export const QueryContext = React.createContext({});
