--- client/src/contexts/product-context.js
@@ -1,7 +1,0 @@
-import React from 'react';
-
-export const ProductContext = React.createContext(null);
-
-export const QueryContext = React.createContext(null);
-
-export const AllProductsContext = React.createContext(null);
